Add option to force creating symlinks.

+++ create_symlinks.py
@@ -2,6 +2,7 @@
 # Makes symlinks in $HOME for all the files in this directory (except this one)
 from __future__ import print_function
 import os
+import sys
 
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -23,13 +24,16 @@ def main():
     for file in os.listdir(THIS_DIR):
         source = os.path.join(THIS_DIR, file)
         symlink = os.path.join(HOME, file)
+        force = '-f' in sys.argv
         if not exclude(source):
-            try:
+            if not os.path.exists(symlink) or force:
+                if os.path.exists(symlink):
+                    os.remove(symlink)
                 os.symlink(source, symlink)
                 print('ln -s {} {}'.format(source, symlink))
-            except OSError as e:
-                if e.errno == 17:
-                    print('file {} already exists'.format(symlink))
+            else:
+                print('file {} already exists'.format(symlink))
+
 
 
 if __name__ == '__main__':
